backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.10.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.0
httpx==0.27.0
